Titre de votre PR

+++ nester.py
@@ -4,6 +4,7 @@ from flask_sqlalchemy import SQLAlchemy
 import time
 import threading
 import datetime
+import socket
 
 app = Flask(__name__) # Instance de la classe Flask 
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///bdd.db' #URI de la bdd qui va être crée  
@@ -33,11 +34,11 @@ class HarvesterDetails(db.Model):
     agent_version = db.Column(db.String(50))
 
 # Création de la forme de la db 
-def create_tables():
-    db.create_all()
+#def create_tables():
+#   db.create_all()
 
-def setup():
-    create_tables()
+#def setup():
+#   create_tables()
                     
 def manage_status_v2():
     with app.app_context():
@@ -49,28 +50,19 @@ def manage_status_v2():
                     search_id = client.id
                     print(f"[NESTER][INFO] Checking the status of : {client.hostname}")
                     time.sleep(3)
-                    if now - float(client.last_request) > 60 and now - float(client.last_request) < 350:
+                    if now - float(client.last_request) > 20 and now - float(client.last_request) < 350:
                         print(now)
                         print(float(client.last_request))
                         print(now - float(client.last_request))
                         client.statut = "Disconnected"
-                    elif now - float(client.last_request) > 600:
+                    elif now - float(client.last_request) > 600: # Cette partie du code peut rendre une erreur en essayant de détruire une entité déjà inexistante
                         del_client = db.session.query(Data).filter_by(id=client.id).first()
-                        db.session.delete(del_client)
+                        if del_client is not None: # Vérification du statut de l'entité avant .delete 
+                            db.session.delete(del_client)
             except Exception as e: 
                 print(f"[NESTER][STATUT_MANAGER][ERROR] {e}")
             db.session.commit()
-            
-def auto_delete_hostname():
-    with app.app_context():
-        while True: 
-            all_client = Data.query.all()
-            
-                
 
-# Lancement d'un thread en parallèle de l'exécution de Flask    
-# check_statut = threading.Thread(target=manage_status_v2)
-# check_statut.start()
 
 # Pour utiliser cette partie, executer : client.py ou le stack de client avec 'docker-compose up' 
 # Cette route n'accepte que les requêtes PUT (Création et mise à jour)
@@ -140,7 +132,7 @@ def client_details():
             agent_version = data_details.get('agent_version')
         )
         db.session.add(new_data_details)
-    
+        
     db.session.commit()
     timestamp = datetime.datetime.now().strftime("%d/%m/%Y %H:%M:%S")
     return jsonify({
@@ -163,7 +155,7 @@ def connexion():
         else:
             flash('Connexion refusée. Merci de réessayer votre mot de passe !')
             print("Connexion échouée")
-    return render_template('login.html') # Si la requète est "GET" (récupération de donnée )
+    return render_template('login.html', hostname=socket.gethostname()) # Si la requète est "GET" (récupération de donnée )
 
 
 #Voir les infos des clients
@@ -177,7 +169,7 @@ def view_client_info():
 def details(hostname):
     data_details = HarvesterDetails.query.filter_by(hostname=hostname).first()
     if data_details: 
-        return render_template('detail.html', data_details=data_details)
+        return render_template('clientdb.html', data_details=data_details)
     else: 
         return "Aucune information trouvée pour le hostname spécifié.", 404
